Rename old IngressRouteTCP module
