+++ tembo-operator/src/ingress.rs
@@ -1,4 +1,4 @@
-use crate::ingress_route_tcp_crd::{
+use crate::ingress_route_tcp_crd_old::{
     IngressRouteTCP, IngressRouteTCPRoutes, IngressRouteTCPRoutesMiddlewares, IngressRouteTCPRoutesServices,
     IngressRouteTCPSpec, IngressRouteTCPTls,
 };
