+++ tembo-operator/src/traefik/mod.rs
@@ -1,4 +1,4 @@
 pub mod ingress_route_crd;
-pub mod ingress_route_tcp_crd;
+pub mod ingress_route_tcp_crd_old;
 pub mod middleware_tcp_crd;
 pub mod middlewares_crd;
